LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-navigation-navigate-expected.txt


FAIL navigate event for navigation.navigate() assert_equals: expected (object) null but got (undefined) undefined
